Add Dockerfile for easier development
This is for those that can't/don't want to install OpenMPI on their
machine.
The docker container will start an SSH daemon on the default port 22,
you can bind this port to wherever you want.
Public keys are fetched from github.
The `start_docker` will do the binding for you if you want.